Remove need for no_cpu
